server/requirements.txt
# Core framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Database
sqlalchemy[asyncio]==2.0.48
asyncpg==0.31.0
alembic==1.18.4

# Auth
python-jose[cryptography]==3.5.0
passlib==1.7.4
bcrypt==4.2.1
pydantic[email]==2.11.9
email-validator==2.3.0

# External APIs
assemblyai==0.54.0
groq==1.0.0
langfuse==3.14.5
